requirements.txt
pyautogui==0.9.54
opencv-python==4.8.1.78
pytesseract==0.3.10
google-generativeai==0.3.2
pynput==1.7.6
Pillow==10.0.1
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
psutil==5.9.6
